examples/dapp/pyproject.toml
[tool.poetry]
name = "dapp"
version = "0.1.0"
description = ""
authors = ["Predrag Peranović <predrag.peranovic@blokkod.me>"]

[tool.poetry.dependencies]
jembe = {path="../.."}
python = "^3.8"
python-dotenv = "^0.14.0"
Flask-SQLAlchemy = "^2.4.4"

[tool.poetry.dev-dependencies]
mypy = "^0.782"
black = "^19.10b0"
pytest = "^6.0.1"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
